requirements.txt
bottle==0.12.13
future==0.16.0
oauthlib==2.0.2
requests-oauthlib==0.8.0
python-twitter==3.3
requests==2.14.2